Introduce `directory` feature module
Allow payjoin and payjoin-directory to share ShortId code.
