+++ payjoin-directory/Cargo.toml
@@ -26,6 +26,7 @@ hyper = { version = "1", features = ["http1", "server"] }
 hyper-rustls = { version = "0.26", optional = true }
 hyper-util = { version = "0.1", features = ["tokio"] }
 ohttp = { package = "bitcoin-ohttp", version = "0.6.0"}
+payjoin = { version = "0.22.0", features = ["directory"], default-features = false }
 redis = { version = "0.23.3", features = ["aio", "tokio-comp"] }
 rustls = { version = "0.22.4", optional = true }
 tokio = { version = "1.12.0", features = ["full"] }
