+++ payjoin/src/uri/mod.rs
@@ -4,6 +4,8 @@ use bitcoin::address::NetworkChecked;
 pub use error::PjParseError;
 use url::Url;
 
+#[cfg(feature = "v2")]
+pub(crate) use crate::directory::ShortId;
 use crate::uri::error::InternalPjParseError;
 #[cfg(feature = "v2")]
 pub(crate) use crate::uri::url_ext::UrlExt;
@@ -11,64 +13,6 @@ pub(crate) use crate::uri::url_ext::UrlExt;
 pub mod error;
 #[cfg(feature = "v2")]
 pub(crate) mod url_ext;
-
-#[cfg(feature = "v2")]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ShortId(pub [u8; 8]);
-
-#[cfg(feature = "v2")]
-impl ShortId {
-    pub fn as_bytes(&self) -> &[u8] { &self.0 }
-    pub fn as_slice(&self) -> &[u8] { &self.0 }
-}
-
-#[cfg(feature = "v2")]
-impl std::fmt::Display for ShortId {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        let id_hrp = bitcoin::bech32::Hrp::parse("ID").unwrap();
-        f.write_str(
-            crate::bech32::nochecksum::encode(id_hrp, &self.0)
-                .expect("bech32 encoding of short ID must succeed")
-                .strip_prefix("ID1")
-                .expect("human readable part must be ID1"),
-        )
-    }
-}
-
-#[cfg(feature = "v2")]
-#[derive(Debug)]
-pub enum ShortIdError {
-    DecodeBech32(bitcoin::bech32::primitives::decode::CheckedHrpstringError),
-    IncorrectLength(std::array::TryFromSliceError),
-}
-
-#[cfg(feature = "v2")]
-impl std::convert::From<bitcoin::hashes::sha256::Hash> for ShortId {
-    fn from(h: bitcoin::hashes::sha256::Hash) -> Self {
-        bitcoin::hashes::Hash::as_byte_array(&h)[..8]
-            .try_into()
-            .expect("truncating SHA256 to 8 bytes should always succeed")
-    }
-}
-
-#[cfg(feature = "v2")]
-impl std::convert::TryFrom<&[u8]> for ShortId {
-    type Error = ShortIdError;
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        let bytes: [u8; 8] = bytes.try_into().map_err(ShortIdError::IncorrectLength)?;
-        Ok(Self(bytes))
-    }
-}
-
-#[cfg(feature = "v2")]
-impl std::str::FromStr for ShortId {
-    type Err = ShortIdError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let (_, bytes) = crate::bech32::nochecksum::decode(&("ID1".to_string() + s))
-            .map_err(ShortIdError::DecodeBech32)?;
-        (&bytes[..]).try_into()
-    }
-}
 
 #[derive(Debug, Clone)]
 pub enum MaybePayjoinExtras {
